LayoutTests/fast/images/image-formats-support.html

<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<style>
    img {
        width: 100px;
        height: 100px;
    }
</style>
</head>
<body>
    <img id="image-1" src="">
    <img id="image-2" src="">
    <img id="image-3" src="">
    <img id="image-4" src="">
    <script>
        // The test finishes after getting the results of loading all the images.
        jsTestIsAsync = true;

        description("Test the whitelist of the image formats in WebKit.");

        var images = [
            { id: "image-1", src: "resources/100x100-red.psd", shouldBeSupported: false },
            { id: "image-2", src: "resources/100x100-red.tga", shouldBeSupported: false },
            { id: "image-3", src: "resources/100x100-red-psd-renamed.png", shouldBeSupported: false },
            { id: "image-4", src: "resources/100x100-red-tga-renamed.png", shouldBeSupported: false },
        ];

        var finishedImages = 0;

        function trimImageSource(src) {
            var index = src.lastIndexOf('/');
            if (index != -1)
                index = src.lastIndexOf('/', index - 1);
            return index != -1 ? src.slice(index + 1) : src;
        }

        function shouldBeSupported(src) {
            var image = images.find(function(image) {
                return image.src == src;
            });

            if (image == undefined) {
                debug("Error: Unknown image src.");
                return false;
            }

            return image.shouldBeSupported;
        }

        function onImageFinish(src, isSupported) {
            shouldBe("shouldBeSupported('" + trimImageSource(src) + "').toString()", "'" + isSupported + "'");
            if (++finishedImages == images.length)
                finishJSTest();
        }

        function onImageLoad(event) {
            onImageFinish(document.getElementById(event.target.id).src, true);
        }

        function onImageError() {
            onImageFinish(document.getElementById(event.target.id).src, false);
        }

        images.forEach(function(image) {
            var element = document.getElementById(image.id);
            element.onload = onImageLoad;
            element.onerror = onImageError;
            document.getElementById(image.id).src = image.src;
        });
    </script>
    <script src="../../resources/js-test-post.js"></script>
</body>
</html>
